requirements.txt
pyzmq
pycapnp
ipython
